Use concurrently to manage startup and termination of tests
